+++ apps/io/nocom/dlist-final/write/NullArrayOutputStream.java
@@ -3,10 +3,10 @@ import ibis.io.ArrayOutputStream;
 
 final class NullArrayOutputStream extends ArrayOutputStream { 
 	
-	int len = 0;
+	long len = 0;
 
 	public int getAndReset() { 
-		int temp = len;
+		int temp = (int) len;
 		len = 0;
 		return temp;
 	} 
@@ -57,7 +57,7 @@ final class NullArrayOutputStream extends ArrayOutputStream {
 	}
 
 	public long bytesWritten() { 
-		return (long) len;
+		return len;
 	} 
 
 	public void resetBytesWritten() {
