+++ apps/io/nocom/object-array/write/NullArrayOutputStream.java
@@ -4,10 +4,10 @@ import ibis.io.ArrayOutputStream;
 
 final class NullArrayOutputStream extends ArrayOutputStream { 
 	
-	int len = 0;
+	long len = 0;
 
 	public int getAndReset() { 
-		int temp = len;
+		int temp = (int) len;
 		len = 0;
 		return temp;
 	} 
@@ -62,6 +62,6 @@ final class NullArrayOutputStream extends ArrayOutputStream {
 	}
 
 	public long bytesWritten() { 
-		return (long) len;
+		return len;
 	} 
 }
